Update vCenter big VM overrides
We want to keep DRS enabled and set it to "partiallyAutomated". Since
we've already added some overrides, we have to handle the case, where
the override already exists and has a different value.

Additionally, this commit fixes the usage of `ReconfigureCluster_Task`
which is deprecated since 2.5. One should use `ConfigSpecEx` and
`ReconfigureComputeResource_Task` instead.