[vcenter] handle backing uuid mismatch properly and less wrong logging
there was still one case where the vcenter consistency nanny had a
false positive: when the backing uuid of the volume was wrong, then
it was assuming the volume to be not mounted - this is fixed now by
adding the uuid extracted from the filename of the volume file to
the list of mounted volumes as well.

besides that all new inconsistency logging has been removed in the
consistency tool interactive mode where we do not need it.