[nova-consistency] ignore block_device_mappings with null volume_id
There are cases when an boot-from-volume instance is being created,
so nova first creates the BDMs in the DB before creating a volume.
Once the volume is created, nova will update the BDM with volume_id.
Meanwhile, nanny might have already mark that BDM as deleted since
the volume_id was None and it was not found in the volume list.

This issue was also seen in the logs as follows:
`block device mapping (id in nova db: <id>) for non existent volume in cinder: None`