nova-instance-mapping: Fix always match first cell
When finding the cell for an instance a cursor is rightfully created
for running the query, but when checking the rowcount to see if any
instances matched, a new cursor was created. New cursors still have
rowcount > 0 (presumably some query is run when creating the cursor).
This made the first checked cell always match and put all instances
with NULL cell_mapping into the same first cell, regardless of the
actual cell they're in.

Create the cursor separately and use it for query and rowcount-check.