[manila] fix check affinity
The rules are defined in metadata of a share (m), and it contains a list
of shares that should be in affinity or anti-affinity relationship with
the share 'm'. The old checks were too strong, assuming the shares in
rule are also in affinity to each other. They are actually not subject
to such relationships.