[manila] snapshot quota usage: only count once per snapshot (#86)
That reflects how manila is determining snapshot quota on creating
a snapshot. Snapshot instances belonging to non-active share replicas
are not taken into account.

The quota sync was created before the replication feature was
introduced. That is why this was not considered probably.

We are okay with not setting quota usage for non-active snapshots,
because the quota usage is anyhow a lot higher than physical used
storage space by snapshots in most cases.